backend/requirements.txt
Flask
flask-cors
python-docx
